+++ mosakin-backend/src/main/kotlin/com/mosa/office/kintai/application/model/Slack.kt
@@ -1,0 +1,7 @@
+package com.mosa.office.kintai.application.model
+
+data class Slack (
+        var text: String?
+)
+
+class SlackMessageException : Exception()
