+++ mosakin-backend/src/main/kotlin/com/mosa/office/kintai/application/usecase/AddPaidUseCase.kt
@@ -1,6 +1,7 @@
 package com.mosa.office.kintai.application.usecase
 
 import com.mosa.office.kintai.application.service.CurrentUserService
+import com.mosa.office.kintai.application.service.SlackService
 import com.mosa.office.kintai.application.service.UniqueIdGenerator
 import com.mosa.office.kintai.application.transaction.TransactionBoundary
 import com.mosa.office.kintai.domain.model.Paid
@@ -8,7 +9,6 @@ import com.mosa.office.kintai.domain.model.PaidTimeType
 import com.mosa.office.kintai.domain.service.PaidService
 import org.springframework.stereotype.Component
 import java.time.LocalDate
-import com.mosa.office.kintai.domain.service.SlackService
 
 /**
  * UseCase層はアプリケーションとしての要求をDomainServiceやDomainModel、Repositoryを利用して実現します。
@@ -39,7 +39,7 @@ class AddPaidUseCase(
         }
         // TODO Slackへの通知
         // TODO 送る文字列の作成
-        slackService.postSlackMessage(" \n新規： *" + "ユーザー名" + "*\n日時： *" + input.paidAcquisitionDate + " (" +  ")" + "  [" + input.paidTimeType + "]*\n```\n" + input.paidReason + "```")
+        slackService.postAddSlackMessage(input)
     }
 
 }
