+++ mosakin-backend/src/main/kotlin/com/mosa/office/kintai/controller/AddPaidController.kt
@@ -3,7 +3,7 @@ package com.mosa.office.kintai.controller
 import com.mosa.office.kintai.application.usecase.AddPaidInputDto
 import com.mosa.office.kintai.application.usecase.AddPaidUseCase
 import com.mosa.office.kintai.domain.model.DuplicatedPaidException
-import com.mosa.office.kintai.domain.model.SlackMessageException
+import com.mosa.office.kintai.application.model.SlackMessageException
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
@@ -17,11 +17,12 @@ class AddPaidController(private val useCase: AddPaidUseCase) {
         try {
             useCase.add(input)
         } catch(e:Exception) {
+            println(e)
             return when(e){
                 // TODO ここら辺はうまくException Handlerとかに逃がしたい
                 is DuplicatedPaidException -> "君の登録しようとしてる有給は他の有給と日付が被っているよ！"
-                is SlackMessageException -> "slackのURLが設定されていないから送れませんでした〜"
-                else -> "謎のエラーだよ"
+                is SlackMessageException -> "slackが送れないよ!"
+                else -> throw e
             }
         }
         return "追加したよ！"
