--- mosakin-backend/src/main/kotlin/com/mosa/office/kintai/domain/model/Slack.kt
@@ -1,10 +1,0 @@
-package com.mosa.office.kintai.domain.model
-
-data class Slack(
-        val url : String
-)
-data class Text (
-        var text: String?
-)
-
-class SlackMessageException : Exception()
