--- mosakin-backend/src/main/kotlin/com/mosa/office/kintai/domain/service/SlackService.kt
@@ -1,32 +1,0 @@
-package com.mosa.office.kintai.domain.service
-
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
-import com.fasterxml.jackson.module.kotlin.readValue
-import com.mosa.office.kintai.config.SlackConfig
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.http.MediaType
-import org.springframework.http.RequestEntity
-import org.springframework.stereotype.Service
-import org.springframework.web.client.RestTemplate
-import java.io.InputStream
-import java.net.URI
-import com.mosa.office.kintai.domain.model.Slack
-import com.mosa.office.kintai.domain.model.Text
-import com.mosa.office.kintai.domain.model.SlackMessageException
-
-@Service
-class SlackService(
-) {
-    @Autowired
-    lateinit var slackConfig: SlackConfig
-
-    fun postSlackMessage(message: String?) {
-        val text: Text = Text(message);
-        val mapper = jacksonObjectMapper()
-        val jsonData:String = slackConfig.slack() ?: return throw SlackMessageException()
-        val data = mapper.readValue<Slack>(jsonData)
-        val request: RequestEntity<Text> = RequestEntity.post(URI(data.url)).accept(MediaType.APPLICATION_JSON).body(text);
-        val restTemplate = RestTemplate()
-        restTemplate.exchange(request, String::class.java)
-    }
-}
